:bug: Fix false positive when looking for value return
The existing code was seeing a return of a value for a function if it
contained a function (or contained a function that contained a function,
or...) that returned a value.

This change seeks to ensure that we never look inside sub-functions to
decide if the current function returns something.